==> 04-missing.Rmd <==
# Missing values

Below are some key observations with respect to Missing Values in the dataset. 

```{r}
library(tidyverse)
library(patchwork)
library(ggplot2)
library(mi)
library(extracat)
library(dplyr)
```

```{r}
theme_fakeNews_rotate <- function(base_size = 14, ang=90) {
  theme_bw(base_size = base_size) %+replace%
    theme(
      # L'ensemble de la figure
      plot.title = element_text(size = 25, face = "bold", margin = margin(10,10,10,10), hjust = 0.5),
      # Zone où se situe le graphique
      panel.grid.minor = element_blank(),
      panel.border = element_blank(),
      # Les axes
      axis.title = element_text(size = 13, face = "bold", margin = margin(10,10,10,10)),
      axis.text.x = element_text(size = 13, face = "bold", margin = margin(10,10,20,10), angle = ang),
      axis.text.y = element_text(size = 13, face="bold", margin = margin(10,10,10,10)),
      axis.line = element_line(color = "black", arrow = arrow(length = unit(0.3, "lines"), type = "closed")),
      # La légende
      legend.title = element_text(size = 13, face = "bold"),
      legend.text = element_text(size = 13, face = "bold")
    )}

theme_fakeNews <- function(base_size = 14) {
  theme_bw(base_size = base_size) %+replace%
    theme(
      # L'ensemble de la figure
      plot.title = element_text(size = 25, face = "bold", margin = margin(10,10,10,10), hjust = 0.5),
      # Zone où se situe le graphique
      panel.grid.minor = element_blank(),
      panel.border = element_blank(),
      # Les axes
      axis.title = element_text(size = 13, face = "bold", margin = margin(10,10,10,10)),
      axis.text.x = element_text(size = 13, face = "bold", margin = margin(10,10,10,10)),
      axis.text.y = element_text(size = 13, face="bold", margin = margin(10,10,10,10)),
      axis.line = element_line(color = "black", arrow = arrow(length = unit(0.3, "lines"), type = "closed")),
      # La légende
      legend.title = element_text(size = 13, face = "bold"),
      legend.text = element_text(size = 13, face = "bold")
    )}
```

```{r}
missingGraphPropotion <- function(df, percent=FALSE, shorten=FALSE) {
  
  # initializing variables depending on percentage or count
  alph <- 0.4
  p1_left <- 'num'
  p2_bot <- 'row count'
  limit <- NULL
  
  if(percent){
    p1_left <- '%'
    p2_bot <- '% rows'
    limit <- c(0,100)
  }
  sh <- waiver()
  if(shorten){
    sh <- abbreviate
  }

  # gets null proportion per column
  hor.index <- function(df) {
    v <- c()
    for (i in head(names(df), -1)) {
      if(percent){
        frac <- 100*sum(df$count[c(df[,i]==TRUE)])/sum(df$count)
      } else {
        frac <- sum(df$count[c(df[,i]==TRUE)])
      }
      v <- c(v, frac)
    }
    data <- data.frame(
      name = head(names(df), -1),
      value = v
    )
    return(data)
  }
  
  # organizes dataframe by null pattern, and uses the null porportion to reorder columns
  null.graph <- function(df) {
    ms <- data.frame(is.na(df)) %>%
    group_by_all() %>%
    dplyr::count(name = "count", sort = TRUE) %>%
    ungroup()
    
    ct <- ms$count
    ind <- hor.index(ms)$value
    ms <- ms[, order(ind, decreasing = TRUE)]
    ms$count = ct
    return(ms)
  }
  
  g <- null.graph(df)
  ind <- hor.index(g)
  
  
  # top side plot
  p1 <- ggplot(data = ind, aes(x = reorder(name, desc(value)), y = value)) + 
    geom_bar(color='white', stat = 'identity', alpha = alph) +
    ggtitle('Missing value patterns') +
    ylab(paste(p1_left, 'rows missing:')) +
    xlab('') +
    scale_y_continuous(expand = c(0,0), limits = limit, n.breaks = 5) +
    scale_x_discrete(labels = sh) +
    theme(
      # L'ensemble de la figure
      plot.title = element_text(size = 25, face = "bold", margin = margin(10,10,10,10), hjust = 0.5),
      # Zone où se situe le graphique
      panel.grid.minor = element_blank(),
      panel.border = element_blank(),
      # Les axes
      axis.title = element_text(size = 13, face = "bold", margin = margin(10,10,10,10)),
      axis.text.x = element_text(size = 13, face = "bold", margin = margin(10,10,10,10), angle = 90),
      axis.text.y = element_text(size = 13, face="bold", margin = margin(10,10,10,10)),
      axis.line = element_line(color = "black", arrow = arrow(length = unit(0.3, "lines"), type = "closed")),
      # La légende
      legend.title = element_text(size = 13, face = "bold"),
      legend.text = element_text(size = 13, face = "bold"))
  
  if(percent){
    g$count <- 100*g$count/sum(g$count)
  }

  
  # make pl dataframe for the main plot
  g$row=as.numeric(row.names(g))
  pl <- pivot_longer(data = g, 
               cols = -c(count, row),
               names_to = "name",
               values_to = "vals")
  pl <- merge(pl, ind)
  
  
  # make fill values for alpha
  g2 <- g[1:(length(g)-2)]
  empty <- c()
  for (row in 1:nrow(g2)) {
    empty <- c(empty, sum(g2[row,]))
  }
  fill <- data.frame(
    row = c(1:nrow(g2)),
    fill = empty
  )
  
  
  # merge fill values to pl, and add fill column to g
  pl <- merge(pl, fill)
  g$fill <- fill$fill
  
  # if there are no missing columns
  if(sum(g$fill)==0){
    alph = 1
  }
  
  # right side plot
  p2 <- ggplot(data = g, aes(x = reorder(as.numeric(row.names(g)), desc(as.numeric(row.names(g)))), y = count, alpha = fill == 0)) + 
    geom_bar(color='white', stat = 'identity') +
    ylab(p2_bot) +
    xlab('') +
    scale_y_continuous(expand = c(0,0), limits = limit, n.breaks = 5) +
    scale_alpha_manual(values = c(alph, 1)) +
    coord_flip() +
    theme(
      # L'ensemble de la figure
      plot.title = element_text(size = 25, face = "bold", margin = margin(10,10,10,10), hjust = 0.5),
      # Zone où se situe le graphique
      panel.grid.minor = element_blank(),
      panel.border = element_blank(),
      # Les axes
      axis.title = element_text(size = 13, face = "bold", margin = margin(10,10,10,10)),
      axis.text.x = element_text(size = 13, face = "bold", margin = margin(10,10,10,10)),
      axis.text.y = element_text(size = 13, face="bold", margin = margin(10,10,10,10)),
      axis.line = element_line(color = "black", arrow = arrow(length = unit(0.3, "lines"), type = "closed")),
      # La légende
      legend.title = element_text(size = 13, face = "bold"),
      legend.text = element_text(size = 13, face = "bold"))
  
  
  # main plot
  p3 <- ggplot(data = pl, aes(x = reorder(name, desc(value)), y = reorder(row, desc(row)), fill = vals, alpha = fill == 0)) +
  geom_tile(color = 'white') +
  scale_alpha_manual(values = c(alph, 1)) +
  ggplot2::annotate('text', x = length(x = unique(pl$name))/2 + 0.5, y = tail(pl$row,1) - pl[pl$fill == 0,]$row[1] + 1, label = 'Complete Cases') +
  scale_x_discrete(labels = sh) +
  theme(
      # L'ensemble de la figure
      plot.title = element_text(size = 25, face = "bold", margin = margin(10,10,10,10), hjust = 0.5),
      # Zone où se situe le graphique
      panel.grid.minor = element_blank(),
      panel.border = element_blank(),
      # Les axes
      axis.title = element_text(size = 13, face = "bold", margin = margin(10,10,10,10)),
      axis.text.x = element_text(size = 13, face = "bold", margin = margin(10,10,10,10), angle = 90),
      axis.text.y = element_text(size = 13, face="bold", margin = margin(10,10,10,10)),
      axis.line = element_line(color = "black", arrow = arrow(length = unit(0.3, "lines"), type = "closed")),
      # La légende
      legend.title = element_text(size = 13, face = "bold"),
      legend.text = element_text(size = 13, face = "bold")) +
  xlab('variable') +
  ylab('missing pattern')
  
  
  # combine plots with patchwork
  pp <- p1 + plot_spacer() + p3 + p2 +
    plot_layout(widths = c(5,1), heights = c(1,4))
  
  return(pp)
}
```



```{r}
df_missing = read_csv(file = 'FakeNews_Unclean.csv' )
```



**Missing values by column for the Liar dataset**

```{r}
colSums(is.na(df_missing)) %>%
  sort(decreasing = TRUE)
```


1) It is important to note that the Label column (independent feature) doesn't have missing values in the dataset. 

2) Most of the columns have 2 missing values, which after taking a closer look belongs to ID's 5872 and 8180

3) Speaker Job Title has the most missing values - 3170 missing entries


**Top 30 row-id having maximum missing values**

```{r}
temp_df = df_missing
row.names(temp_df) <- 1 : 11378
t = rowSums(is.na(temp_df)) %>%
  sort(decreasing = TRUE)

head(t,30)
```


1) Maximum number of columns with missing values in a given row are 11. For examples Id's like 5872, 8180 etc. have 11 columns with missing values.


```{r echo=FALSE, message=FALSE, results='hide'} 

#Ratio of Missing Values for each Label in Test dataset 

features = c("Label", "Speaker Job Title")
df_new = df_missing[features]

df_new  = df_missing %>%
group_by(Label) %>%
summarise_each(funs(sum(is.na(.))/length(.)))

df_new1 = df_new[order(df_new$`Speaker Job Title`, decreasing=TRUE),]

barplot(df_new1$`Speaker Job Title`*100 , 
        names.arg = df_new1$Label,
        xlab = "Label of fake news",
        ylab = "Percentage of Missing Values",
        main ="Percentage of missing values with respect to Label") + theme_fakeNews()

```


**Heatmap **

```{r fig.width=10}
df_pivot <- df_missing %>%
    gather(key, value, -Label) %>% 
    mutate(missing = ifelse(is.na(value), "yes", "no"))

ggplot(df_pivot, aes(y = key, x =  fct_rev(Label), fill = missing)) +
  geom_tile(color = "white") + 
  labs(title = "Missing Values for Features Faceted by Label",
       x = "Label Associated", y="Features in the Dataset") + theme_fakeNews()
```



Since we noticed that "Speaker Job Title" and "State Info" had maximum missing values, we used heatmaps to see if we can observe any pattern. Below are the observed patterns :-

  1) For "Speaker Job Title", True & False Staements  were completely missing.
  
  2) For "State Info" feature, False Statements were completely missing. 

**Missing proportions**

```{r fig.width=17, fig.height=12}
missingGraphPropotion(df_missing) 
```

We created this missing values plots to identify any patterns for the rows. 

  1) The majority of the rows have no missing values
  2) Roughly 1800 rows have Speaker Job Title and State Info missing
  3) Over 1000 rows only have Speaker Job Title missing, and roughly 500 rows only have State Info missing.
  4) The rest of the patterns are hardly present.
