==> 06-interactive.Rmd <==
# Interactive component


```{r}

#fakenews = read.csv(file = 'DataFakeNews.csv')


#Removing Chain Email from list of speakers
#fakenews = fakenews[!(fakenews$Speaker == "chain-email"),]
```



```{r}
#Taking the speakers with the top 10 fake statements
#wow = fakenews %>% distinct(Speaker,.keep_all = TRUE) %>%  group_by(The.Party.Affiliation)
#pow = wow %>% summarise(Speaker = Speaker, p_bt = sum(Barely.True.Counts), p_fc = sum(False.Counts), p_ht = #sum(Half.True.Counts), p_mt = sum(Mostly.True.Counts), p_pfic = sum(Pants.on.Fire.Counts))
#fakenews_1 = fakenews %>% mutate(Ts = rowSums(.[10:14]))
#final = inner_join(pow,fakenews_1,by = c("Speaker","The.Party.Affiliation"))

```

For the Interactive Portion, we have **tried building a game** for a user, where he sees a statement and **guess if it's True, False, Barely-True, Mostly-Ture, Pants On Fire.** 

Step 1) The user is expected to select a speaker from the list of top 10 Speaker (filtered by the number of statements in csv) displayed below.

Step 2) The user sees a dynamic graph for the Party Affiliation of the Speaker, and its history in spreading Fake Vs Factual News. **Please note**, the user is expected to click on the statement where The Party Name is Shown. 

Step 3) Then user sees the history for the speaker in terms of Fake Vs Factual statements made by the speaker in the past. **Please note**, the user is again expected to click on the statement for the next step. 

Step 4) Now the user is presented with a **random statment** made by the speaker at a particular location. The user has to judge whether the statement is True, Mostly True, Barely True,  Half True, False, or Pants on Fire using the dropdown list given. The user is alerted as to whether he chose the correct option and the game resets.

User can continue to play endless games in the given above order.



<iframe src="Demo.html" width="600" height="600"></iframe>
